requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pandas

requests

apache-airflow-providers-postgres

apache-airflow-providers-microsoft-azure

azure-identity

apache-airflow-providers-slack

apache-airflow-providers-http